+++ packages/python-ta/CHANGELOG.md
@@ -20,6 +20,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Updated `invalid_name_checker.py` to include a suggested fix for invalid names in checks using snake_case format.
 - Updated `SnapshotTracer` to convert snapshots to JSON with `snapshot_to_json` and pass JSON data to the webstepper template instead of SVG.
 - Added configuration options for maximum name lengths of each name category in `invalid_name_checker.py`.
+- Update configuration options to support `.toml` files
 
 ### 💫 New checkers